+++ odk-connector/models/__init__.py
@@ -3,3 +3,5 @@
 from . import odk
 from . import odk_config_model
 from . import odk_submissions_model
+from . import odk_config_service
+from . import odk_submissions_service
